+++ src/index.ts
@@ -1,4 +1,6 @@
 export * from "./git";
+export * from "./github";
 export * from "./markdown";
 export * from "./config";
 export * from "./semver";
+export * from "./repo";