+++ src/markdown.ts
@@ -3,6 +3,7 @@ import { convert } from "convert-gitmoji";
 import { fetch } from "node-fetch-native";
 import type { ChangelogConfig } from "./config";
 import type { GitCommit, Reference } from "./git";
+import { formatReference, formatCompareChanges } from "./repo";
 
 export async function generateMarkDown(
   commits: GitCommit[],
@@ -17,13 +18,8 @@ export async function generateMarkDown(
   const v = config.newVersion && `v${config.newVersion}`;
   markdown.push("", "## " + (v || `${config.from}...${config.to}`), "");
 
-  if (config.github) {
-    markdown.push(
-      `[compare changes](https://github.com/${config.github}/compare/${
-        config.from
-      }...${v || config.to})`,
-      ""
-    );
+  if (config.repo) {
+    markdown.push(formatCompareChanges(v, config), "");
   }
 
   for (const type in config.types) {
@@ -68,7 +64,7 @@ export async function generateMarkDown(
     [..._authors.keys()].map(async (authorName) => {
       const meta = _authors.get(authorName);
       for (const email of meta.email) {
-        const { user } = await fetch(`https://ungh.unjs.io/user/find/${email}`)
+        const { user } = await fetch(`https://ungh.cc/users/find/${email}`)
           .then((r) => r.json())
           .catch(() => ({ user: null }));
         if (user) {
@@ -102,6 +98,34 @@ export async function generateMarkDown(
   return convert(markdown.join("\n").trim(), true);
 }
 
+export function parseChangelogMarkdown(contents: string) {
+  const headings = [...contents.matchAll(CHANGELOG_RELEASE_HEAD_RE)];
+  const releases: { version?: string; body: string }[] = [];
+
+  for (let i = 0; i < headings.length; i++) {
+    const heading = headings[i];
+    const nextHeading = headings[i + 1];
+    const [, title] = heading;
+    const version = title.match(VERSION_RE);
+    const release = {
+      version: version ? version[1] : undefined,
+      body: contents
+        .slice(
+          heading.index + heading[0].length,
+          nextHeading?.index ?? contents.length
+        )
+        .trim(),
+    };
+    releases.push(release);
+  }
+
+  return {
+    releases,
+  };
+}
+
+// --- Internal utils ---
+
 function formatCommit(commit: GitCommit, config: ChangelogConfig) {
   return (
     "  - " +
@@ -112,33 +136,20 @@ function formatCommit(commit: GitCommit, config: ChangelogConfig) {
   );
 }
 
-const refTypeMap: Record<Reference["type"], string> = {
-  "pull-request": "pull",
-  hash: "commit",
-  issue: "ssue",
-};
-
-function formatReference(ref: Reference, config: ChangelogConfig) {
-  if (!config.github) {
-    return ref.value;
-  }
-  return `[${ref.value}](https://github.com/${config.github}/${
-    refTypeMap[ref.type]
-  }/${ref.value.replace(/^#/, "")})`;
-}
-
 function formatReferences(references: Reference[], config: ChangelogConfig) {
   const pr = references.filter((ref) => ref.type === "pull-request");
   const issue = references.filter((ref) => ref.type === "issue");
   if (pr.length > 0 || issue.length > 0) {
     return (
       " (" +
-      [...pr, ...issue].map((ref) => formatReference(ref, config)).join(", ") +
+      [...pr, ...issue]
+        .map((ref) => formatReference(ref, config.repo))
+        .join(", ") +
       ")"
     );
   }
   if (references.length > 0) {
-    return " (" + formatReference(references[0], config) + ")";
+    return " (" + formatReference(references[0], config.repo) + ")";
   }
   return "";
 }
@@ -162,3 +173,6 @@ function groupBy(items: any[], key: string) {
   }
   return groups;
 }
+
+const CHANGELOG_RELEASE_HEAD_RE = /^#{2,}\s+.*(v?(\d+\.\d+\.\d+)).*$/gm;
+const VERSION_RE = /^v?(\d+\.\d+\.\d+)$/;